refactor(backend): use bennu's person find endpoint instead
The latest version of tecnico-api has not been deployed yet,
which prevents us to use its person find endpoint.

Therefore, we have to use bennu's endpoint, which searches everyone
instead of only students of a certain degree, but on the other hand
allows users to search by IST-ID.

This commit should be reverted once tecnico-api is deployed.

For now, this means #85 can be postponed to a later release.